This is an edited version of count_worms_directory.m, which I decided to keep as-is for now (as it is intended for the original protocol, which we do not use). This version supports 2- and 4-spot assays, and introduces the following naming scheme:
[Sorting index]_[Condition name]_[Spot name]_p[N]_processed.tiff

Where:
Sorting index is a number (using a constant format of three digits should be best) that both specifies the location of the worm count in the resulting output table and serves to aggregate the different spots from the same plate into the same row.
Condition name is a free-form string (spaces allowed) that names the condition (e.g., "N2 Mock But -3").
Spot name is a free-form string (spaces allowed) that identifies the spot (e.g., "But-3", "Condition", "Control"), and also used to name the columns in the resulting output table (and aggregate the worm counts under them). Each plate should have the same number and name for spots.
_p[N] is an optional specifier (that is, it doesn't have to be in the name at all) with N being the number of worms to add over and above what is counted in the picture. Can be useful for when the worms spread out over a large area and some don't enter the frame.
_proccessed is a mandatory (for now) suffix that was historically added since the original pictures were darkfield, which the script cannot handle, and so we have a separate ImageJ macro that converts them to brightfield.